requirements.txt
django
django-cors-headers
djangochannelsrestframework